fix: moved the generate metadata function into one file
