+++ src/app/[netname]/layout.tsx
@@ -14,15 +14,15 @@ const NETNAME_MAP: { [key in Netname]: string } = {
   degen: 'Degen Chain',
 };
 
-export async function generateMetadata({ params }: Props): Promise<Metadata> {
-  const netname = NETNAME_MAP[params.netname] || 'Base';
+// export async function generateMetadata({ params }: Props): Promise<Metadata> {
+//   const netname = NETNAME_MAP[params.netname] || 'Base';
 
-  return {
-    title: `poidh on ${netname} - pics or it didn't happen`,
-    description:
-      "poidh - pics or it didn't happen - fully onchain bounties + collectible NFTs - start your collection today on Arbitrum, Base, or Degen Chain",
-  };
-}
+//   return {
+//     title: `poidh on ${netname} - pics or it didn't happen`,
+//     description:
+//       "poidh - pics or it didn't happen - fully onchain bounties + collectible NFTs - start your collection today on Arbitrum, Base, or Degen Chain",
+//   };
+// }
 
 export default function Layout({ children }: Props) {
   return children;
